[plugin.video.si] 3.0.2
