+++ plugin.video.si/changelog.txt
@@ -1,3 +1,4 @@
+version 3.0.2 website changes
 version 3.0.1 Isengard version - separate scraper
 version 1.4.5 cleanup of brightcove and website changes
 version 1.4.3 fixed website changes
