+++ plugin.video.si/default.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # KodiAddon 
 #
-from lib.scraper import myAddon
+from resources.lib.scraper import myAddon
 import re
 import sys
 
@@ -10,5 +10,4 @@ import sys
 addonName = re.search('plugin\://plugin.video.(.+?)/',str(sys.argv[0])).group(1)
 ma = myAddon(addonName)
 ma.processAddonEvent()
-del myAddon
