+++ plugin.video.si/resources/__init__.py
