--- plugin.video.si/resources/language/English/strings.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<strings>
-    <string id="30003">Swim Daily</string>
-
-    <string id="30020">View</string>
-    <string id="30021">Enable Views</string>
-    <string id="30022">Default View</string>
-    <string id="30023">Movies View</string>
-    <string id="30024">Shows View</string>
-    <string id="30025">Episodes View</string>
-
-</strings>
